update mkvlib:fix bugs

+++ mkvlib/mkv.go
@@ -312,7 +312,7 @@ func (self *mkvProcessor) ASSFontSubset(files []string, fonts, output string, di
 	obj.fonts = findFonts(obj._fonts)
 	obj.loadCache(self.cache)
 	r := obj.parse() && obj.matchFonts() && obj.createFontsSubset() && obj.changeFontsName() && obj.replaceFontNameInAss()
-	if self.a2p {
+	if r && self.a2p {
 		r = self.ass2Pgs(obj._files, self.pr, self.pf, obj.output, d, lcb)
 		if r && !self.apc {
 			_ = os.RemoveAll(obj.output)
@@ -345,7 +345,7 @@ func (self *mkvProcessor) GetFontsList(input string, lcb logCallback) []string {
 	return nil
 }
 
-func (self *mkvProcessor) CreateFontsCache(dir, output string, lcb logCallback) bool {
+func (self *mkvProcessor) CreateFontsCache(dir, output string, lcb logCallback) []string {
 	obj := new(assProcessor)
 	obj._fonts = dir
 	obj.lcb = lcb
